+++ History.md
@@ -1,3 +1,5 @@
+* RobMen: Add TouchFile custom action.
+
 ## WixBuild: Version 4.0.2102.0
 
 * RobMen: Merge recent changes through WiX v3.9.901.0